feat: 평점api 구현중

+++ src/main/java/com/example/knockknock/entity/Rate.java
@@ -20,7 +20,7 @@ public class Rate extends BaseTimeEntity {
     private Long id;
 
     @Column(name = "value", nullable = false)
-    private int value;
+    private float value;
 
     @Column(name = "content", nullable = false)
     private String content;
